Did some more vision testing. Adjusted the AutonDriveFunctionsTest program to attempt to use vision to check ring stack and then move accordingly.
Things to check:
Amount of time it has to detect the stack might be the barrier from it detecting.
Also might be an issue with the setup itself, although it mimics previously working vision classes very closely.
Likely that both of the these issues in tandem are causing the bad.